A contribution-based client selection method.
We study the client selection problem with differences in client value. Define the contribution of a single client as the influence on the global model accuracy, and propose a client selection algorithm (CBE3) based on the contribution. It updates the clients selection weights by estimating the expectation of the impact of different clients on the accuracy of the system global model. The CBE3 strategy is highly adaptive, and the adjustment coefficients can achieve a balance between global model accuracy and convergence speed.